move message response to a strategy class
