+++ Raft/Election/IMessageResponseStrategy.cs
@@ -1,0 +1,9 @@
+using Raft.Entities;
+
+namespace Raft.Election
+{
+    public interface IMessageResponseStrategy
+    {
+        void RespondToMessage(NodeMessage message);
+    }
+}
